[Style]: 코드 포맷 수정, applicationState 코드 병합
