+++ presentation/src/main/kotlin/ac/dnd/bookkeeping/android/presentation/ui/bottombar/BottomBarItem.kt
@@ -1,7 +1,7 @@
 package ac.dnd.bookkeeping.android.presentation.ui.bottombar
 
 import ac.dnd.bookkeeping.android.presentation.R
-import ac.dnd.bookkeeping.android.presentation.common.root.MainRoot
+import ac.dnd.bookkeeping.android.presentation.common.root.MainRootConstant
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
 
@@ -12,35 +12,35 @@ sealed class BottomBarItem(
 ) {
     data object BottomFirst :
         BottomBarItem(
-            route = MainRoot.BOTTOM_FIRST,
+            route = MainRootConstant.BOTTOM_FIRST,
             drawableResId = R.drawable.bottom_sample,
             stringResId = R.string.bottom_sample
         )
 
     data object BottomSecond :
         BottomBarItem(
-            route = MainRoot.BOTTOM_SECOND,
+            route = MainRootConstant.BOTTOM_SECOND,
             drawableResId = R.drawable.bottom_sample,
             stringResId = R.string.bottom_sample
         )
 
     data object BottomThird :
         BottomBarItem(
-            route = MainRoot.BOTTOM_THIRD,
+            route = MainRootConstant.BOTTOM_THIRD,
             drawableResId = R.drawable.bottom_sample,
             stringResId = R.string.bottom_sample
         )
 
     data object BottomFourth :
         BottomBarItem(
-            route = MainRoot.BOTTOM_FOURTH,
+            route = MainRootConstant.BOTTOM_FOURTH,
             drawableResId = R.drawable.bottom_sample,
             stringResId = R.string.bottom_sample
         )
 
     data object BottomFifth :
         BottomBarItem(
-            route = MainRoot.BOTTOM_FIFTH,
+            route = MainRootConstant.BOTTOM_FIFTH,
             drawableResId = R.drawable.bottom_sample,
             stringResId = R.string.bottom_sample
         )
